Move LocalFeedImageLoader to production scope
